+++ NEWS.md
@@ -4,6 +4,7 @@
 
 * Introduces `wait_aio()`, a user-interruptible version of `call_aio()`.
 * Implements `%~>%` signal forwarder from one 'conditionVariable' to another.
+
 #### Updates
 
 * `until()` updated to be identical to `.until()`, returning FALSE instead of TRUE if the timeout has been reached.